Added Custom Exceptions

+++ src/main/java/com/javaproject/javaspring/controllers/EmployeeController.java
@@ -14,6 +14,9 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
+import com.javaproject.javaspring.exception.BadRequestException;
+import com.javaproject.javaspring.exception.EmployeeExistException;
+import com.javaproject.javaspring.exception.EmployeeNotFound;
 import com.javaproject.javaspring.models.Employee;
 import com.javaproject.javaspring.repo.EmployeeRepo;
 
@@ -36,32 +39,55 @@ public class EmployeeController {
     }
 
     @GetMapping(value = "/getEmployee/{empId}")
-    public Employee getEmployee(@PathVariable int empId) {
-        return employeeRepo.findById(empId).orElseThrow(() -> new IllegalStateException("Employee with id " + empId + " does not exist"));
+    public Employee getEmployee(@PathVariable int empId) throws EmployeeNotFound, BadRequestException{
+        if (empId <= 0){
+            throw new BadRequestException("Please use a valid employee ID");
+        }
+        return employeeRepo.findById(empId).orElseThrow(() -> new EmployeeNotFound("Employee with id " + empId + " does not exist"));
     }
 
     @PostMapping(value="/addEmployee")
-    public  ResponseEntity<String> addEmployee(@Validated @NonNull @RequestBody Employee newEmployee) {
+    public  ResponseEntity<String> addEmployee( @Validated @NonNull @RequestBody Employee newEmployee) throws EmployeeExistException, EmployeeNotFound, BadRequestException {
+        if (newEmployee == null){
+            throw new BadRequestException("Please have a valid request body");
+        }
+        List<Employee> employees = employeeRepo.findAll();
+        for (Employee e: employees){
+            if(e.getEmpId() == newEmployee.getEmpId()){
+                throw new EmployeeExistException("This employee already exists");
+            }
+        }
         employeeRepo.save(newEmployee);
         return new ResponseEntity<String>("Employee has successfully been added!", HttpStatus.CREATED);
     }
     
     @PutMapping(value="/updateEmployee/{empId}")
-    public ResponseEntity<String> updateEmployee(@PathVariable int empId, @Validated @NonNull @RequestBody Employee updateEmployee) {
-
-        Employee employee = employeeRepo.findById(empId).orElseThrow(() -> new IllegalStateException("Employee with id " + empId + " does not exist"));
-        employee.setEmpName(updateEmployee.getEmpName());
-        employee.setEmpLastName(updateEmployee.getEmpLastName());
-        employee.setCellNumber(updateEmployee.getCellNumber());
-        employee.setEmail(updateEmployee.getEmail());
-        employee.setRole(updateEmployee.getRole());
-        employee.setSalary(updateEmployee.getSalary());
-        employeeRepo.save(employee);
+    public ResponseEntity<String> updateEmployee(@PathVariable int empId, @Validated @NonNull @RequestBody Employee updateEmployee) throws EmployeeNotFound, BadRequestException {
+        if (empId <= 0){
+            throw new BadRequestException("Please enter a valid employee ID");
+        }
+        if (updateEmployee == null){
+            throw new BadRequestException("Please have a valid request body");
+        }
+        Employee employee = employeeRepo.findById(empId).orElseThrow(() -> new EmployeeNotFound("Employee with id " + empId + " does not exist"));
+        if (employee != null){
+            employee.setEmpNumber(updateEmployee.getEmployeeNumber());
+            employee.setEmpName(updateEmployee.getEmpName());
+            employee.setEmpLastName(updateEmployee.getEmpLastName());
+            employee.setCellNumber(updateEmployee.getCellNumber());
+            employee.setEmail(updateEmployee.getEmail());
+            employee.setRole(updateEmployee.getRole());
+            employee.setSalary(updateEmployee.getSalary());
+            employeeRepo.save(employee);
+        }
         return new ResponseEntity<>("Employee with id " + empId +" has been updated", HttpStatus.OK);
     }
 
     @DeleteMapping(value="/deleteEmployee/{empId}")
-    public ResponseEntity<String> deleteEmployee(@PathVariable int empId) {
+    public ResponseEntity<String> deleteEmployee(@PathVariable int empId) throws BadRequestException{
+        if (empId <= 0){
+            throw new BadRequestException("Please enter a valid employee ID");
+        }
         employeeRepo.deleteById(empId);
         return new ResponseEntity<String>("Employee with id " + empId + " has been deleted", HttpStatus.OK);
     }
